+++ src/main/java/org/rs2server/rs2/model/skills/Cooking.java
@@ -174,7 +174,7 @@ public class Cooking extends ProductionAction {
 		}
 
 		public int getEndXp() {
-			return (int) endXp;
+			return endXp;
 		}
 
 		public Item getBurntProduct() {
