+++ src/main/java/org/rs2server/rs2/model/skills/fish/Fish.java
@@ -59,7 +59,7 @@ public enum Fish {
 	}
 
 	public int getXp() {
-		return (int) xp;
+		return xp;
 	}
 
 	public int getLevel() {
